Allow lambda functions to be a part of a function sequence.
Allow multiply statements in a single line, without using a semicolon as
a separator. This doesn't always works, so if in doubt a semicolon should
be used. This change on semantics, will assist in a more natural code flow,
because sometimes it makes sence like:

 lambda (x) { if (x > 10) { return x / 2 } return x * 2} (10)

though it is still a bit hard to parse when you read it, but it is natural
when you write it. But multiply statements in a signle line, it is a bad
programming practice anyway.

+++ src/z/la-modules/file/file-module.c
@@ -354,7 +354,7 @@ static VALUE file_writelines (la_t *this, VALUE v_file, VALUE v_ar) {
   string **ar = (string **) AS_ARRAY(array->value);
 
   for (size_t i = 0; i < array->len; i++) {
-    if (fprintf (fp, "%s\n", ar[i]->bytes) isnot ar[i]->num_bytes + 1) {
+    if (fprintf (fp, "%s\n", ar[i]->bytes) isnot (int) ar[i]->num_bytes + 1) {
       fclose (fp);
       return INT(LA_NOTOK);
     }
